+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+### Added
+ - Option to force add reply button
+
 ## 1.1.6 - 2020-08-07
 ### Fix
  - Timestamps to not create extra line